+++ src/viewers/spacetimegraphdata.h
@@ -30,7 +30,7 @@ public:
     Q_ENUM(SpaceTimeType)
     SpaceTimeGraphData(SpaceTimeType type, Direction direction = Direction::dFull);
     ~SpaceTimeGraphData();
-    GraphDataset getDataset(KLP::FrameCollection const& collection, int sliceIndex) override;
+    GraphDataset getDataset(KLP::FrameCollection const& collection, qint64 sliceIndex) override;
     int type() const override { return mType; }
 
 private:
